Put the z3-related functions in a separate file in order to minimize modifications
